Turn anonymous class into a named class
So that the buildDispatchers method look a little manageable
